authentication/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic
SQLAlchemy
mysql
mysql-connector-python-rf
mysql-connector-python
python-multipart
